move air_storage func to utils
