+++ .github/scripts/utils.py
@@ -1,5 +1,7 @@
 import os
 import re
+import copy
+import string
 import yaml
 import json
 import shutil
@@ -276,7 +278,6 @@ def load_price_data(file_path: str = "./src/data/cars_dealer_price.json") -> Dic
         return {}
 
 
-# Пример использования:
 def update_car_prices(car, prices_data: Dict[str, Dict[str, int]]) -> None:
     """
     Обновляет цены в XML элементе автомобиля.
@@ -611,3 +612,88 @@ def update_yaml(car, filename, friendly_url, current_thumbs, config):
 
     return filename
 
+
+# Создаем последовательность a-z + 0-9
+chars = string.ascii_lowercase + string.digits
+base = len(chars)  # Основание для системы исчисления (36)
+
+def vin_to_number(vin):
+    """Конвертирует последние цифры VIN в число."""
+    if not vin[-5:].isdigit():
+        raise ValueError("Последние 5 символов VIN должны быть цифрами.")
+    
+    return int(vin[-5:])  # Преобразуем последние 5 символов VIN в число
+
+def number_to_vin(vin, number):
+    """Преобразует число обратно в VIN."""
+    new_suffix = str(number).zfill(5)  # Преобразуем число обратно в строку с ведущими нулями
+    return vin[:-5] + new_suffix  # Собираем новый VIN
+
+def modify_vin(vin, increment):
+    """Изменяет VIN путем увеличения последних цифр."""
+    vin_number = vin_to_number(vin)  # Получаем числовое значение последних 5 цифр VIN
+    new_vin_number = vin_number + increment  # Увеличиваем на заданное значение
+    return number_to_vin(vin, new_vin_number)  # Преобразуем обратно в VIN
+
+def str_to_base36(str):
+    """Конвертирует строку STR в число на основе системы с основанием 36."""
+    value = 0
+    for char in str:
+        value = value * base + chars.index(char)  # Преобразуем каждый символ в число
+    return value
+
+def base36_to_str(value, length):
+    """Конвертирует число обратно в строку STR на основе системы с основанием 36."""
+    str = []
+    while value > 0:
+        str.append(chars[value % base])
+        value //= base
+    return ''.join(reversed(str)).zfill(length)  # Добавляем нули в начало, если нужно
+
+def increment_str(str, increment):
+    """Изменяет STR путем увеличения всей строки на значение increment."""
+    str_value = str_to_base36(str)  # Конвертируем STRING в число
+    new_str_value = str_value + increment  # Увеличиваем на заданное значение
+    return base36_to_str(new_str_value, len(str))  # Преобразуем обратно в строку
+
+def duplicate_car(car, config, n, status="в пути", offset=0):
+    """Функция для дублирования элемента 'car' N раз с изменением vin."""
+    duplicates = []
+
+    # Проверка наличия обязательных полей 'VIN' и 'Availability'
+    try:
+        if car.find(config['vin_tag']) is None:
+            raise ValueError(f"Элемент 'car' не содержит обязательного поля '${config['vin_tag']}'")
+        if car.find(config['availability_tag']) is None:
+            raise ValueError(f"Элемент 'car' не содержит обязательного поля '${config['availability_tag']}'")
+    except ValueError as e:
+        print(f"Ошибка: {e}")
+        return duplicates  # Вернем пустой список и продолжим выполнение скрипта
+    
+    for i in range(n):
+        try:
+            new_car = copy.deepcopy(car)  # Клонируем текущий элемент car
+            
+            # Обрабатываем VIN
+            vin = new_car.find(config['vin_tag']).text
+            new_vin = modify_vin(vin.lower(), offset+i+1)
+            new_car.find(config['vin_tag']).text = new_vin.upper()  # Меняем текст VIN
+            
+            # Обрабатываем unique_id, если он существует
+            unique_id_element = new_car.find(config['unique_id_tag'])
+            if unique_id_element is not None:
+                unique_id = unique_id_element.text
+                new_unique_id = increment_str(unique_id, offset + i + 1)  # Изменяем последний символ на i
+                unique_id_element.text = new_unique_id  # Меняем текст unique_id
+                print(vin, new_vin, unique_id, new_unique_id)
+            else:
+                print(vin, new_vin, f"${config['unique_id_tag']} отсутствует")
+            
+            # Обновляем статус
+            new_car.find(config['availability_tag']).text = status  # Меняем статус Наличие автомобиля
+            duplicates.append(new_car)
+        
+        except AttributeError as e:
+            print(f"Ошибка при обработке элемента: {e}")
+    
+    return duplicates
